+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Feature - data cleaning history api
 ## Current Release 
 ### 0.208.0 
 **Release Date:** Tue Aug 28 00:34:21 UTC 2018